requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyJWT==2.8.0
bcrypt==4.1.3
mysql-connector-python==9.0.0
blueprint==3.4.2
python-dotenv==1.0.1
timedelta==2020.12.3
